+++ trunk/gxlib/xatom.h
@@ -70,7 +70,7 @@ private:
     LockedAtomSphere, // 1 list
     ConstrainedAtomSphere // 1 list
     ;
-  static double MinQAlpha;
+  static float MinQAlpha;
   Poly* Polyhedron;
   TXGlLabel* Label;
   bool label_forced;
@@ -153,8 +153,8 @@ public:
   static void SetQPeakScale(float V);    // to use with q-peaks
   static float GetQPeakSizeScale();    // to use with q-peaks
   static void SetQPeakSizeScale(float V);    // to use with q-peaks
-  static double GetMinQAlpha() { return MinQAlpha; }
-  static void SetMinQAlpha(double v) { MinQAlpha = v; }
+  static float GetMinQAlpha() { return MinQAlpha; }
+  static void SetMinQAlpha(float v) { MinQAlpha = v; }
   void CalcRad(short DefAtomR);
 
   void ApplyStyle(TGraphicsStyle& S);
